Fixed a problem of multiple IOCs on same host
By default, IOC will use port 5064 for both UDP and TCP connections.
This worked great for PVmapper as it assumed both are the same. But
when multiple IOC are running on the same host, only the first one will
use TCP port 5064 and the others will not. The TCP port is included in
the UDP response but PVmapper ignored it and used the UDP port instead.