Set IOC guard's non-blocking option after sending initial ECHO
Part of the IOC guard's initialization is to send an ECHO packet.
Previously this was done after socket has been set to non-blocking and
it could fail. Rather than making the mechanism of determining if IOC
is connected a lot more complicated, let's try to change the socket to
non-blocking AFTER sending the initial ECHO. This will block foir a
while but hopefully it doesn't cause to much of a slow-down.