Optimized removing PV from other searchers when PV is found.
It appears that the approach with a global variable was slowing
things down considerably. Instead of switching to try using other
containers for the job (std::map), decided to rework the mechanism
and remove the PV when it is actually found. This way, we search
each list exactly once when PV is found, instead of every time
we're constructing the list of PVs to be searched for.