Change from single search interval to a list of intervals
For more flexibility and customizing how and when the PVs are searched
for, PVmapper now let's the site configure multiple search intervals.
For each retry, the next search interval from the list is used, until
list is exhausted and the last interval is then used until PV is found.

While adding this functionality, we needed the Searcher::scheduleNextSearch()
to accurately insert into the sorted list, which unfortunately it did
not. So in order to fix it, we added basic unit testing for it. And
while at it, we simplify processOutgoing() function and optimize
processIncoming() function.

Can't wait to test it all in a real 1-mio-PVs
environment tomorrow.