Balance the searches for constant search rate
Instead of one search queue sorted by next-search timestamp,
keep a list of search buckets, one bucket for each 0.1s search.
Every time the purge runs, re-balance the buckets to be about
the same size. This allows to search for missing PVs at a
constant rate, rather than based on when the PVs were originally
searched for, which might be common with big screens.