Balancing was too aggresive for small number of PVs and was creating more UDP traffic
The balancing was trying to make every bin about equal size. But
multiple PVs can fit in a single UPD packet. For a small number of
constantly searching PVs, the balancing actually increased number
of UDP requests. This patch tries to squash up to 10 bins into
one, which means not all bins are the same size but number of
UDP requests is decreased up to 10 times.